Add files via upload
Merged concat_fasta.py with prune_tree.py. When I was testing, the only problem with the pruning step is that I was using the old version of the fasta file: all species' names (a total of 108) were adapted to the old master tree file, in which some species were named differently compared to the new 202 taxa master tree. Like Old: GORILLA_GORILLA was changed to GORILLA_GORILLA_GORILLA in the 202 taxa tree file, so it didn't match and got trimmed out. Please test it on your end to ensure it works and the pruned file matches with the actual overlapping taxa.